@jkwatson 2021-05-28
I went ahead and ran the SpanPipelineBenchmark before & after this change. (note: this benchmark was broken due to drift in the collector config. also, I made it run 100 spans per iteration, rather than just a single one, to make it slightly more realistic).

Before this change:
```
Benchmark                                                                                                Mode  Cnt       Score       Error   Units
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline                                   thrpt   15    6641.008 ±   380.803   ops/s
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.alloc.rate                    thrpt   15    1167.901 ±    69.683  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.alloc.rate.norm               thrpt   15  277078.863 ±  4942.548    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Eden_Space           thrpt   15    1166.414 ±    98.418  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Eden_Space.norm      thrpt   15  276961.958 ± 21179.001    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Old_Gen              thrpt   15       0.412 ±     0.229  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Old_Gen.norm         thrpt   15      97.393 ±    52.417    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Survivor_Space       thrpt   15       0.885 ±     0.743  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Survivor_Space.norm  thrpt   15     207.630 ±   169.578    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.count                         thrpt   15      98.000              counts
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.time                          thrpt   15     231.000                  ms
```

After this change:
```
Benchmark                                                                                                Mode  Cnt       Score       Error   Units
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline                                   thrpt   15    6593.079 ±   178.843   ops/s
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.alloc.rate                    thrpt   15    1170.622 ±    31.610  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.alloc.rate.norm               thrpt   15  279727.981 ±   929.051    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Eden_Space           thrpt   15    1172.692 ±    64.946  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Eden_Space.norm      thrpt   15  280262.959 ± 14679.709    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Old_Gen              thrpt   15       0.539 ±     0.314  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Old_Gen.norm         thrpt   15     128.827 ±    75.708    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Survivor_Space       thrpt   15       0.707 ±     0.780  MB/sec
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.churn.G1_Survivor_Space.norm  thrpt   15     169.759 ±   186.648    B/op
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.count                         thrpt   15     114.000              counts
SpanPipelineBenchmark.BatchSpanProcessorBenchmark.measureSpanPipeline:·gc.time                          thrpt   15     285.000                  ms
```

I can't see from this benchmark that it makes any measurable difference. Am I missing something?