@jkwatson 2021-05-28
I think this saves us about 80 bytes per span in the OTLP pipeline, if my math is correct.