@anuraaga 2021-05-29
Yup that's my math too. I don't think it's actually the answer to @iNikem's woes but is a trivial microwin, which after several does add up so let's take it.